feat: split api methods into modules
