+++ src/api/invoices.js
@@ -1,0 +1,5 @@
+export default {
+  getInvoices() {
+    return this.$fetch('GET', '/invoices');
+  }
+};
